+++ directio-project/asakusa-directio-dmdl/src/main/resources/com/asakusafw/dmdl/directio/util/messages.properties
@@ -1,0 +1,12 @@
+AttributeAnalyzer.diagnosticMapKeyNotCharacter=key of {0} must be a single character
+AttributeAnalyzer.diagnosticMapValueNotCharacter=value of {0} must be a single character
+AttributeAnalyzer.diagnosticNotBoolean={0} must be boolean value
+AttributeAnalyzer.diagnosticNotCharacter={0} must be a single character
+AttributeAnalyzer.diagnosticNotCharsetName={0} must be a charset name
+AttributeAnalyzer.diagnosticNotClassName={0} must be a class name
+AttributeAnalyzer.diagnosticNotDateFormat={0} must be a valid pattern string of "SimpleDateFormat"
+AttributeAnalyzer.diagnosticNotDecimalFormat={0} must be a valid pattern string of "DecimalFormat"
+AttributeAnalyzer.diagnosticNotEnumConstant={0} must be one of '{'{1}'}'
+AttributeAnalyzer.diagnosticNotMap={0} must be a map
+AttributeAnalyzer.diagnosticNotString={0} must be a string
+AttributeAnalyzer.diagnosticNotTimeZoneId={0} must be a valid time zone ID
